pyproject.toml
[project]
name = "mailalfred"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "beautifulsoup4>=4.13.0",
    "chromadb>=1.5.2",
    "duckdb>=1.4.4",
    "google-api-python-client>=2.150.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "httpx>=0.28.0",
    "jinja2>=3.1.6",
    "loguru>=0.7.3",
    "openai-agents>=0.10.4",
    "python-dotenv>=1.2.2",
    "pytz>=2026.1.post1",
    "tqdm>=4.67.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
